replace existing connections table & rework what 'warning' means
